added file_handling file to edit the file in a separate class
